Implemented basic settings support.
Following the tradition set forth by Django, uppercase items in the settings
module are assumed to be attributes set by the user.  This allows us to import
the settings file, and iterate over the elements, ignoring those that are not
uppercase.

firmant/wsgi.py was changed to print out a list of the set attributes in the
settings file.

settings.py has a dummy setting to show the syntax.

firmant/settings.py contains the settings class.  It is a loose wrapper around a
dictionary and takes as its constructor a string referencing the module to
import.