Removed all resolver code.
It is currently too early to make that design decision.

I know the design pattern I wish to follow is to have each URL map to a
function, and have that function take a request and keywords to return a
response.  I will develop such functions and then implement the URL
mapping (possibly using code deleted in this commit).