Refactored the settings code.
Among other things, it now uses a configure once, import everywhere
approach.  This means that the WSGI interface only has to initialize
settings once, and then they stay initialized wherever they were
imported.

It also remains coherent.   If one were to import the settings in module
A, and reconfigure in module B, then the changes are reflected in module
A as well.