Added the atom schema tests.
Currently it only tests creating the schema and dropping the schema.
Both of these actions are important to ensuring that our schema is
correct with respect to PostgreSQL syntax.

Also included in this commit is a fix to test/settings.py which makes it
so all settings operations in the TestSettings class do not interfere
with the other unit test's settings.