Moved some code into the while loop.
It turns out that in the Relation class, if the result set is 0 tuples,
it would call len(None) which would throw an exception.  Rather than
catching the exception, I just moved the call to len inside a loop where
the conditional guarantees that row will not be none.  This is slightly
inefficient but works for now (len is called on each iteration but only
needs to be checked once).