Removed the connection factory from Relation.
It did not enhance the Relation class or its subclasses, and I can
envision needing to bypass it anyway in the near future and generate
more than one type of connection from the same object.  Overall, very
little is lost by removing this, and it removes a potential source for
confusion if people are wondering where to get their connection from.