Fixed integration between postgresql and pytz.
For example:
In America/New_York, April 3, 2009 is during daylight savings time.  If
Postgresql were queried for a timestamp with such a date and timezone,
it would return "2009-04-03 21:31:30".  It would return the timezone as
'EDT' using the to_char function.

Pytz does not recognize 'EDT' as a valid timezone.  This patch should
fix that.