Switched the backend to the new plugins system.
The firmant/backend/atom.py was updated to contain an AtomProvider
plugin mount.  The firmant.frontend modules were updated to use the
AtomProvider for instead of directly importing Entry objects.  Tests
were modified to match.

As part of the new system, tests will be improved upon in the near
future.

The settings file now refers to the ATOM_PROVIDER and not an
ATOM_MODULE.