Removed firmant.plugins.datasource.postgresql
The reasons behind this are plentiful.  Despite being functional and
adhering to the interface it defined, the interface it defined needs to
be changed.  The changes needed are plentiful (for instance Author and
Category should not be namespaced with an underscore, but should be
objects themselves), and the desire to do it slim.  The changes
implemented that require this change in the interface are for the better
(flatfile, stored in an SCM).

Once the task at hand is settled, this code can come back and undergo a
cleanup.  The schema especially is of high quality and should be re-used
in any future implementation effort.