Added SingleProviderPlugin.
In an attempt to avoid import black magic, the PluginMount meta class is
being replaced with explicit plugin lists in the configuration file.

For the ViewProvider, a separate list of enabled plugins was provided on
top of the plugin list.  This is the first in a series of steps designed
to reduce import black magic and make the system more robust.