Removed to_xml from each atom object.
The to_xml test code essentially duplicated the function.  This is a
sign of bad design.  Additionally, the atom objects need not know about
the proper xml format for atom feeds.  Even though they are derived from
the atom specification, they are to be kept as agnostic as possible as
to the manner in which they will be displayed.