Looser coupling.
The AtomFlatfile*Provider classes depended upon calling each other.
They were made to call Atom*Provider proxies instead so that the code
works in the case when multiple modules are used for the four atom
providers.

The tests were updated to call via the proxies so that the proxies are
tested.  The proxies were then updated to actually pass calls through.